refactor wire app
put react/redux wiring into utils
