+++ assets/wire/components/Preview.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { gettext } from '../utils';
+import { gettext } from 'utils';
 
 function Preview({item}) {
     return (
