+++ assets/wire/components/SearchResultsInfo.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { gettext } from '../utils';
+import { gettext } from 'utils';
 
 import './SearchResultsInfo.scss';
