+++ karma.conf.js
@@ -14,6 +14,7 @@ module.exports = function(config) {
 
         webpack: {
             module: webpackConfig.module,
+            resolve: webpackConfig.resolve,
         },
 
         webpackMiddleware: {
