+++ package.json
@@ -3,8 +3,7 @@
     "start": "webpack-dev-server --progress --colors --content-base dist",
     "build": "webpack --config webpack.config.js --progress --profile --colors",
     "lint": "eslint assets --ext .js --ext .jsx",
-    "karma": "karma start --single-run",
-    "flow": "flow"
+    "test": "karma start --single-run"
   },
   "dependencies": {
     "babel-core": "^6.26.0",
@@ -12,7 +11,6 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "classnames": "^2.2.5",
     "css-loader": "^0.28.5",
@@ -40,7 +38,6 @@
   "devDependencies": {
     "eslint": "^4.6.1",
     "eslint-plugin-react": "^7.3.0",
-    "flow-bin": "^0.53.1",
     "react-devtools": "^2.5.0",
     "redux-devtools": "^3.4.0"
   }
